==> src/02_text_trends/05_pub_figures.Rmd <==
---
title: "Publication Figures"
author: "Brandon L. Kramer"
date: "1/18/2020"
output: html_document
---

```{r, include=FALSE}

rm(list = ls())

for (pkg in c("tidyverse", "igraph", "tidytext", "ggplot2", "DT", "maditr", "RPostgreSQL",
              "grid", "gridExtra", "reshape2", "extrafont", "data.table", 
              #"datamodelr", "DiagrammeR", "networkD3",
              "plotly", "stringi")) {library(pkg, character.only = TRUE)}

```

```{r supp fig 2, fig.width=7, fig.height=4}

# connect to postgresql to get our data
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the users_gh data (table of all github users) 
by_year <- dbGetQuery(conn, "SELECT * FROM pubmed_2021.articles_per_year")
filtered_by_year <- dbGetQuery(conn, "SELECT * FROM pubmed_2021.biomedical_human_per_year")

# disconnect from postgresql database 
dbDisconnect(conn)



```

```{r supp fig 2, fig.width=7, fig.height=4}

# connect to postgresql to get our data
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the users_gh data (table of all github users) 
by_year <- dbGetQuery(conn, "SELECT * FROM pubmed_2021.articles_per_year")
filtered_by_year <- dbGetQuery(conn, "SELECT * FROM pubmed_2021.biomedical_human_per_year")

# disconnect from postgresql database 
dbDisconnect(conn)

by_year <- by_year %>% 
  rename(all_pubmed = count) %>% 
  left_join(filtered_by_year, by = "year") %>% 
  rename(human_only = count) %>% 
  select(year, human_only, all_pubmed) %>% 
  pivot_longer(!year, names_to = "group", values_to = "count")

by_year

ggplot(data = by_year, aes(fill = group, x = year, y = count)) + 
  geom_bar(stat="identity", position="dodge") +
  labs(title = "Supp Fig 2. Comparison of PubMed/MEDLINE Abstracts Used (1990-2020)") + 
  theme(axis.title.x = element_blank(),
        plot.title = element_text(size=11, hjust = 0.2),
        legend.title = element_text(size = 10, hjust = 0.5),
        panel.background = element_rect(fill = "white"),
        legend.position="bottom") +
  ylab("Number of Abstracts") +
  scale_y_continuous(breaks = c(0, 100000, 200000, 300000, 400000, 500000, 600000, 700000, 800000, 900000, 
                                1000000, 1100000, 1200000, 1300000), 
                     labels = c("0", "100K", "200K", "300K", "400K", "500K", "600K", "700K", "800K", "900K", "1.0M", "1.1M", "1.2M", "1.3M")) +
  scale_fill_manual(values=c("grey","black"),
                    labels=c("PubMed (All)", "Biomedical, Human-Only Abstracts (Sample)")) +
  guides(fill=guide_legend(title=""))
```

```{r}
# connect to postgresql to get our data
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the users_gh data (table of all github users) 
top_journals <- dbGetQuery(conn, "SELECT * FROM pubmed_2021.biomedical_human_per_journal LIMIT 25")

# disconnect from postgresql database 
dbDisconnect(conn)

top_journals_cleaned = top_journals %>% 
  select(publication, pub_count) %>% 
  mutate(publication = ifelse(test = str_detect(string = publication, 
         pattern = "\\b(?i)(Proceedings of the National Academy of Sciences of the United States of America)\\b"), yes = "PNAS", no = publication),
         publication = ifelse(test = str_detect(string = publication, 
         pattern = "\\b(?i)(Journal of immunology)\\b"), yes = "Journal of immunology", no = publication),
         publication = ifelse(test = str_detect(string = publication,
         pattern = "\\b(?i)(Science \\(New)\\b"), yes = "Science", no = publication),
         publication = ifelse(test = str_detect(string = publication, 
         pattern = "\\b(?i)(IEEE Engineering in Medicine and Biology Society)\\b"), 
         yes = "IEEE Engineering in Medicine and Biology Society", no = publication),
         publication = ifelse(test = str_detect(string = publication,
         pattern = "\\b(?i)(The Journal of neuroscience)\\b"), yes = "The Journal of neuroscience", no = publication),
         publication = ifelse(test = str_detect(string = publication,
         pattern = "\\b(?i)(Clinical cancer research)\\b"), yes = "Clinical cancer research", no = publication)
         ) 

setwd("~/git/diversity/data/journal_rankings/")
write_csv(top_journals_cleaned, "top_journals_biomed_humans_only.csv")
```

```{r figures 1AB, fig.height=8, fig.width=11.5, echo=FALSE}

#h1_all_set_counts_full <- read_rds("~/git/diversity/data/text_results/h1_results/h1_all_set_counts_full.rds")
#h1_all_set_counts_trends <- read_rds("~/git/diversity/data/text_results/h1_results/h1_all_set_counts_trends.rds")
general_pop_terms <- read_rds("~/git/diversity/data/text_results/h1_results/h1_all_set_prc_trends.rds")
h1_all_set_counts_trends <- read_rds("~/git/diversity/data/text_results/h1_results/h1_all_set_counts_trends.rds")
subset_counts <- read_rds("~/git/diversity/data/text_results/h1_results/h1_all_subset_counts_trends.rds")



custom_variables <- c("prc_ancestry", "prc_cultural", "prc_diversity", 
                      "prc_genetic", "prc_population", "prc_racial", "prc_sexgender", "prc_sexuality")
custom_linetype <- c("solid", "solid",  "solid", "twodash", "dotted", "dashed")

`%notin%` <- Negate(`%in%`)

figure_1A <- h1_all_set_counts_trends %>% 
  rename(n = count) %>% 
  # graphing terms over time 
  filter(term %notin% c("total", "population", "genetic")) %>% 
  pivot_wider(names_from = term, values_from = n) %>% 
  ggplot() + 
  #geom_line(aes(y = n, x = year, colour = term), size = 1.3, stat="identity") + 
  geom_line(aes(y = lifecourse, x = year, colour = "lifecourse"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = ancestry, x = year, colour = "ancestry"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = `sex/gender`, x = year, colour = "sex/gender"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = `diversity (all)`, x = year, colour = "diversity (all)"), 
            size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = socioeconomic, x = year, colour = "socioeconomic"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = disability, x = year, colour = "disability"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = `race/ethnicity`, x = year, colour = "race/ethnicity"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = cultural, x = year, colour = "cultural"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = equity, x = year, colour = "equity"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = sexuality, x = year, colour = "sexuality"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = `diversity (social)`, x = year, colour = "diversity (social)"), 
            size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = minority, x = year, colour = "minority"), size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = migration, x = year, colour = "migration"), size = 0.8, stat="identity", linetype = "solid") +
  theme_minimal() + 
  labs(title = "        Figure 1A. Total Growth of Diversity-Related \n Terminology in PubMed/MEDLINE Abstracts (1990-2020)", 
       color =  "Term") + ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("ancestry", "cultural", "disability", 
                              "diversity (all)","diversity (social)", "equity", 
                              "lifecourse", "migration", "minority",
                              "race/ethnicity","sex/gender", "sexuality","socioeconomic"),
                     values=c("#ff8396","#B7D100",  "#61B329",    
                              "#1b7524", "#C0C0C0", "#9a009a",  
                              "#CC0000", "#005BB7", "#000000",   
                              "#3B9AB2", "#E86F00", "#b19cd9", "#f2be00")) +
  scale_x_continuous(limits = c(1990, 2020),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015, 2020)) 
  #scale_y_continuous(limits = c(0, 105000),
  #                   breaks = c(0, 5000, 10000, 15000, 20000, 25000, 30000, 100000))

figure_1A_log <- figure_1A + 
  ylab("Term Count (Log Scale)")  +
  scale_y_log10() +
  labs(title = "        Figure 1B. Total Growth of Diversity-Related \n Terminology in PubMed/MEDLINE Abstracts (1990-2020)", 
       color =  "Term") + ylab("Term Count") +
  scale_y_log10(breaks = c(0, 10, 100, 1000, 10000, 100000, 1000000),
                labels = c("0", "10", "100", "1000", "10000", "100000", "1000000"))

# graphing terms over time 
figure_1B_pre <- general_pop_terms %>% 
  select(-starts_with("cnt_")) %>% 
  ggplot() + 
  geom_line(aes(y = prc_ancestry, x = year, colour = "ancestry"), 
            size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_lifecourse, x = year, colour = "lifecourse"), 
            size = 0.8, stat="identity", linetype = "solid") + #dotted
  geom_line(aes(y = prc_sexgender, x = year, colour = "sex/gender"), 
            size = 0.8, stat="identity", linetype = "solid") + #dotted
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity (all)"), 
            size = 0.8, stat="identity", linetype = "solid") + #dotted
  geom_line(aes(y = prc_class, x = year, colour = "socioeconomic"), 
            size = 0.8, stat="identity", linetype = "solid") + #longdash
  geom_line(aes(y = prc_disability, x = year, colour = "disability"), 
            size = 0.8, stat="identity", linetype = "solid") + #longdash
  geom_line(aes(y = prc_racial, x = year, colour = "race/ethnicity"), 
            size = 0.8, stat="identity", linetype = "solid") + #longdash
  geom_line(aes(y = prc_cultural, x = year, colour = "cultural"), 
            size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_equity, x = year, colour = "equity"), 
            size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_sexuality, x = year, colour = "sexuality"), 
            size = 0.8, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_soc_diversity, x = year, colour = "diversity (social)"), 
            size = 0.8, stat="identity", linetype = "solid") + #dotdash
  geom_line(aes(y = prc_minority, x = year, colour = "minority"), 
            size = 0.8, stat="identity", linetype = "solid") + #dotdash
  geom_line(aes(y = prc_migration, x = year, colour = "migration"), 
            size = 0.8, stat="identity", linetype = "solid") + #dotdash
  labs(title = "        Figure 1C. Proportional Growth Diversity-Related \n Terminology in PubMed/MEDLINE Abstracts (1990-2020)", color =  "Term") + 
  theme_minimal() + 
  scale_color_manual(labels=c("ancestry", "cultural", "disability", 
                              "diversity (all)","diversity (social)", "equity", 
                              "lifecourse", "migration", "minority",
                              "race/ethnicity","sex/gender", "sexuality","socioeconomic"),
                     values=c("#ff8396","#B7D100",  "#61B329",    
                              "#1b7524", "#C0C0C0", "#9a009a",  
                              "#CC0000", "#005BB7", "#000000",   
                              "#3B9AB2", "#E86F00", "#b19cd9", "#f2be00"))  

figure_1B <- figure_1B_pre +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        legend.position = "none",
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_x_continuous(limits = c(1990, 2020),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015, 2020)) +
  scale_y_continuous(breaks = c(0, 5, 10, 15, 20, 25, 30))

# graphing terms over time 
figure_1B_log <- figure_1B + 
  ylab("Proportion of Abstracts (Log Scale)") +
  theme(axis.title.x = element_blank(), 
        legend.position = "none",
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_y_log10() +
  labs(title = "        Figure 1D. Proportional Growth Diversity-Related \n Terminology in PubMed/MEDLINE Abstracts (1990-2020)", color =  "Term") 

legend <- figure_1A +
  theme(axis.title.x = element_blank(), 
        legend.position = "bottom",
        plot.title = element_text(size=12, hjust = 0.5)) 
  #scale_linetype_manual(values = c('solid','solid','solid','dotted','dotted','dotted','dotdash','dotdash','dotdash',
  #                                 'solid','solid','solid','dotted','dotted','dotted','dotdash','dotdash','dotdash')) 

legend <- cowplot::get_legend(legend)

grid.arrange(figure_1A, figure_1A_log,
             figure_1B, figure_1B_log, 
             legend,
             nrow = 3, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,4), c(5,5)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5,2.5, 0.4)
             )

```

```{r epa_figure_1ab, eval=FALSE, echo=FALSE}
ggplotly(figure_1A)
ggplotly(figure_1A_log)
ggplotly(figure_1B)

# raw count growth 
# diversity
4574/380
# population 
1832/145
# genetic
1191/77
# sex/gender
729/33
# aging
593/28
# cultural
478/38
# race/ethnicity 
342/16
```


```{r figures_2AB, fig.height=4, fig.width=11.5, echo=FALSE}

h2_omb_counts <- read_rds("~/git/diversity/data/text_results/h2_results/h2_all_omb_counts.rds") 
h2_omb_prcs <- read_rds("~/git/diversity/data/text_results/h2_results/h2_all_omb_prcs.rds") 

# graphing us-specific terms over time using census categories 
# https://www.census.gov/topics/population/race/about.html

figure_2A <- h2_omb_counts %>% 
  group_by(term, year) %>% 
  summarise(n = sum(count)) %>% 
  arrange(-n) %>% 
  ggplot() + 
  geom_line(aes(y = n, x = year, colour = term), size = 0.8, stat="identity") + 
  theme_minimal() +
  labs(title = "     Figure 2A. Total Growth of US OMB Terms \n in PubMed/MEDLINE Abstracts (1990-2020)",
       color =  "Term") + 
  ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("asian","black", "diversity (social)",
                              "ethnicity", "hispanic/latinx", "native-american",
                              "pacific-islander","race","white"),
                     values=c("#005BB7", "#E86F00", "#800080", 
                              "#61B329", "#3B9AB2", "#ff8396", 
                              "#000000", "#f2be00", "#CC0000")) +
  scale_x_continuous(limits = c(1990, 2020),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015, 2020))

# graphing terms over time 
figure_2B_pre <- h2_omb_prcs %>% 
  ggplot() + 
  geom_line(aes(y = prc_black, x = year, colour = "black"), size = 0.8, stat="identity") +
  geom_line(aes(y = prc_white, x = year, colour = "white"),size = 0.8, stat="identity") +
  geom_line(aes(y = prc_asian, x = year, colour = "asian"), size = 0.8,stat="identity") +
  geom_line(aes(y = prc_hispanic, x = year, colour = "hispanic/latinx"),size = 0.8, stat="identity") +
  geom_line(aes(y = prc_native_american, x = year, colour = "native-american"), size = 0.8,stat="identity") +
  geom_line(aes(y = prc_pacific_islander, x = year, colour = "pacific-islander"), size = 0.8,stat="identity") +
  geom_line(aes(y = prc_ethnic, x = year, colour = "ethnicity"), size = 0.8,stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race"), size = 0.8, stat="identity") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity (social)"), size = 0.8, stat="identity") +
  labs(title = "    Figure 2B. Proportional Growth of US OMB Terms \nin PubMed/MEDLINE Abstracts (1990-2020)",
       color =  "Term") + 
  theme_minimal() +
  ylab("Proportion of Abstracts") +
  scale_color_manual(labels=c("asian","black", "diversity (social)",
                              "ethnicity", "hispanic/latinx", "native-american",
                              "pacific-islander","race","white"),
                     values=c("#005BB7", "#E86F00", "#800080", 
                              "#61B329", "#3B9AB2", "#ff8396", 
                              "#000000", "#f2be00", "#CC0000")) +
  scale_x_continuous(limits = c(1990, 2020),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015, 2020)) 

figure_2B <- figure_2B_pre +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5),
        plot.title = element_text(size=12, hjust = 0.5))

# legend
# graphing terms over time 
legend_2AB <- figure_2B_pre +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "bottom",
        legend.title = element_text(size = 10, hjust = 0.5),
        plot.title = element_text(size=12, hjust = 0.5)) 

legend_2AB <- cowplot::get_legend(legend_2AB)

grid.arrange(figure_2A, figure_2B, 
             legend_2AB, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             heights = c(2.5, 0.4)
             )


```

```{r figures_3AB, fig.height=4, fig.width=11.5, message = FALSE, results = FALSE, warning = FALSE, echo=FALSE}

h2_omb_prcs <- read_rds("~/git/diversity/data/text_results/h2_results/h2_all_omb_prcs.rds") 
h3_soc_div <- h2_omb_prcs %>% 
  select(year, cnt_diversity, prc_diversity) %>% 
  pivot_longer(!year, names_to = "word", values_to = "n") %>% 
  mutate(term = "diversity (social)") %>% 
  select(term, word, n, year) 

h3_counts <- read_rds("~/git/diversity/data/text_results/h3_results/h3_all_counts.rds")
all_pop_prc_counts <- read_rds("~/git/diversity/data/text_results/h3_results/h3_all_prcs.rds")

h3_counts <- h3_counts %>% 
  mutate(term = str_replace(term, "omb/us-census", "us omb"))

figure_3A <- h3_counts %>%   
  bind_rows(h3_soc_div %>% filter(word == "cnt_diversity")) %>% 
  group_by(term, year) %>% 
  summarise(n = sum(n)) %>%
  ggplot() + 
  geom_line(aes(y = n, x = year, colour = term), size = 0.8, stat="identity") + 
  theme_minimal() +
  labs(title = "     Figure 3A. Total Growth of Population Terms \n in PubMed/MEDLINE Abstracts (1990-2020)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Term Count (in Thousands)") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("all population terms", "continental", "directional", 
                              "diversity (social)", "national", "us omb", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00", "#3B9AB2", "#f2be00", "#000000", "#E86F00",
                              "#00468b", "#ff8396", "#9a009a", "#1b7524")) +
  #scale_y_log10() +
  scale_x_continuous(limits = c(1990, 2020),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015, 2020)) +
  scale_y_continuous(limits = c(0, 110000),
                     breaks = c(0, 10000, 20000, 30000, 40000, 50000, 60000, 70000, 80000, 90000, 100000, 110000),
                     labels = c("0", "10", "20", "30", "40", "50", "60", "70", "80", "90", "100", "110"))

figure_3B <- all_pop_prc_counts %>% 
  select(-total, -n) %>% 
  bind_rows(h3_soc_div %>% 
              filter(word == "prc_diversity") %>% 
              select(term, year, n) %>% 
              rename(percentage = n)) %>% 
  ggplot() + 
  geom_line(aes(y = percentage, x = year, colour = term), size = 0.8, stat="identity") +
  labs(title = "    Figure 3B. Proportional Growth of Population Terms \n in PubMed/MEDLINE Abstracts (1990-2020)",
       color =  "Term") + 
  theme_minimal() +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        legend.box = "vertical", 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("all population terms", "continental", "directional", 
                              "diversity (social)", "national", "us omb", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00", "#3B9AB2", "#f2be00", "#000000", "#E86F00",
                              "#00468b", "#ff8396", "#9a009a", "#1b7524")) +
  scale_x_continuous(limits = c(1990, 2020),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015, 2020)) +
  scale_y_continuous(limits = c(0, 28),
                     breaks = c(0, 5, 10, 15, 20, 25, 30))

legend_3AB <- all_pop_prc_counts %>% 
  bind_rows(h3_soc_div %>% filter(word == "prc_diversity")) %>% 
  ggplot() + 
  geom_line(aes(y = percentage, x = year, colour = term), size = 0.8, stat="identity") +
  labs(title = "    Figure 3B. Proportional Growth of Population Terms \n in PubMed/MEDLINE Abstracts (1990-2020)",
       color =  "Term") + 
  theme_minimal() +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        legend.box = "vertical", 
        legend.position = "bottom",
        legend.title = element_text(size = 10, hjust = 0.5),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("all population terms", "continental", "directional", 
                              "diversity (social)", "national", "us omb", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00", "#3B9AB2", "#f2be00", "#000000", "#E86F00",
                              "#00468b", "#ff8396", "#9a009a", "#1b7524")) +
  scale_x_continuous(limits = c(1990, 2020),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015, 2020))

legend_3AB <- cowplot::get_legend(legend_3AB)

#grid.arrange(general_pop_graph, gen_pop_prc_counts_graph, nrow = 1, ncol = 2)

grid.arrange(figure_3A, figure_3B, 
             legend_3AB, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )
```
