==> src/02_text_trends/06_compare_diversity_measures.R <==


compare_diversity_measures <- function(analysis_timeframe){
  
  #rm(list = ls())
  #analysis_timeframe <- 1990
  
  library("readr")
  library("dplyr") 
  library("stringr")  
  library("tidytext")
  library("widyr")
  library("RPostgreSQL")
  library("data.table")
  library("maditr")
  library("purrr")
  source("~/git/diversity/scripts/diversitizeR.R")
  
  # connect to postgresql to get our data
  conn <- dbConnect(drv = PostgreSQL(), 
                    dbname = "sdad", 
                    host = "10.250.124.195", 
                    port = 5432, 
                    user = Sys.getenv("db_userid"), 
                    password = Sys.getenv("db_pwd"))
  
  # query the users_gh data (table of all github users) 
  pubmed_data <- dbGetQuery(conn, str_c(
    "SELECT DISTINCT(fk_pmid), year, abstract  
    FROM pubmed_2021.biomedical_abstracts 
    WHERE year =", analysis_timeframe, ";"))
  
  # disconnect from postgresql database 
  dbDisconnect(conn)
  
  str_c("Started at: ", Sys.time())
  pubmed_data <- pubmed_data %>% 
    mutate(abstract_raw = abstract, 
           abstract = tolower(abstract)) %>% 
    humanizeR(fk_pmid, abstract) %>% 
    compoundR(abstract) 
  
  pubmed_data <- pubmed_data %>% 
    compoundR_v3(abstract_raw)
  str_c("Finished at: ", Sys.time())
  
  # version_1 is based on co-occurences within the abstract 
  setwd("/project/biocomplexity/sdad/projects_data/ncses/oss/brandon_diversity_data/h1_results_0521/")
  terms_matrix <- read_rds(str_c("h1_diversity_terms_matrix_", analysis_timeframe,".rds")) %>% 
    select(fk_pmid, diversity_cnt, soc_diversity) %>% 
    rename(alldiv = diversity_cnt) 
  pubmed_data <- pubmed_data %>% 
    left_join(terms_matrix, by = "fk_pmid") %>% 
    mutate(socdiv_v1 = replace_na(soc_diversity, 0)) %>% 
    mutate(alldiv = replace_na(alldiv, 0)) %>% 
    select(-soc_diversity) 
  rm(terms_matrix)
  
  # version_2 removed the polysemes 
  setwd("~/git/diversity/data/text_results/h1_results/")
  terms_matrix <- read_rds(str_c("h1_diversity_terms_matrix_", analysis_timeframe,".rds")) %>% 
    select(fk_pmid, socdiv_cnt)
  pubmed_data <- pubmed_data %>% 
    left_join(terms_matrix, by = "fk_pmid") %>% 
    mutate(socdiv_v2 = replace_na(socdiv_cnt, 0)) %>% 
    select(-socdiv_cnt)
  rm(terms_matrix)
  
  # version_3 and version_4 based on more stringent exclusion clause 
  pubmed_v3 <- pubmed_data %>% 
    filter(human > 0 | nonhuman == 0) %>% 
    detect_social_diversity_v1(fk_pmid, abstract_raw, 
                            remove_polysemes = FALSE) %>%  
    rename(socdiv_v3 = soc_diversity) %>% 
    select(fk_pmid, socdiv_v3)
  
  pubmed_v4 <- pubmed_data %>% 
    detect_social_diversity_v1(fk_pmid, abstract_raw, 
                            remove_polysemes = TRUE) %>% 
    rename(socdiv_v4 = soc_diversity) %>% 
    select(fk_pmid, socdiv_v4)
  
  pubmed_data <- pubmed_data %>%
    left_join(pubmed_v3, by = "fk_pmid") %>% 
    left_join(pubmed_v4, by = "fk_pmid") %>% 
    mutate(socdiv_v3 = replace_na(socdiv_v3, 0)) %>% 
    mutate(socdiv_v4 = replace_na(socdiv_v4, 0))
  rm(pubmed_v3, pubmed_v4)
  
  pubmed_v5_6 <- pubmed_data %>%
    filter(nonhuman == 0) %>% 
    detect_social_diversity_v1(fk_pmid, abstract_raw, 
                            remove_polysemes = FALSE) %>% 
    rename(socdiv_v5 = soc_diversity) %>% 
    detect_social_diversity_v1(fk_pmid, abstract_raw, 
                            remove_polysemes = TRUE) %>% 
    rename(socdiv_v6 = soc_diversity) %>% 
    select(fk_pmid, socdiv_v5, socdiv_v6)
  pubmed_data <- pubmed_data %>%
    left_join(pubmed_v5_6, by = "fk_pmid") %>% 
    mutate(socdiv_v5 = replace_na(socdiv_v5, 0)) %>% 
    mutate(socdiv_v6 = replace_na(socdiv_v6, 0))
  rm(pubmed_v5_6)
  
  # add in the more_stringent = TRUE clause and then you can run this on a loop through years 
  
  pubmed_v7 <- pubmed_data %>%
    filter(nonhuman == 0) %>% 
    detect_social_diversity_v2(fk_pmid, abstract_raw) %>% 
    rename(socdiv_v7 = soc_diversity) %>% 
    select(fk_pmid, socdiv_v7)
  pubmed_data <- pubmed_data %>%
    left_join(pubmed_v7, by = "fk_pmid") %>% 
    mutate(socdiv_v7 = replace_na(socdiv_v7, 0))
  rm(pubmed_v7)
  
  pubmed_v8_9 <- pubmed_data %>%
    filter(nonhuman == 0) %>% 
    detect_social_diversity_v3(fk_pmid, abstract_raw, 
                               remove_polysemes = FALSE) %>% 
    rename(socdiv_v8 = soc_diversity) %>% 
    detect_social_diversity_v3(fk_pmid, abstract_raw, 
                               remove_polysemes = TRUE) %>% 
    rename(socdiv_v9 = soc_diversity) %>% 
    select(fk_pmid, socdiv_v8, socdiv_v9)
  pubmed_data <- pubmed_data %>%
    left_join(pubmed_v8_9, by = "fk_pmid") %>% 
    mutate(socdiv_v8 = replace_na(socdiv_v8, 0)) %>% 
    mutate(socdiv_v9 = replace_na(socdiv_v9, 0))
  rm(pubmed_v8_9)
  
  setwd("~/git/diversity/data/text_results/h1_results/")
  write_csv(pubmed_data %>% select(-abstract_raw), 
            str_c("h1_soc_div_comps_",analysis_timeframe,".csv"))
  
}

for (year in 1990:2020) {
  compare_diversity_measures(year)
  print(str_c("Just finished: ", year))
}


#chk <- pubmed_data %>% filter(alldiv == 1)

colSums(pubmed_data %>% select(-fk_pmid, -year, -abstract, -abstract_raw))

setwd("~/git/diversity/data/text_results/h1_results/")
all_diversity_comps <- list.files(pattern="h1_soc_div_comps_*") %>% 
  map_df(~read_csv(.)) %>% filter(alldiv == 1)

#colSums(pubmed_data %>% select(-abstract, -abstract_raw))

colSums(all_diversity_comps %>% select(-fk_pmid, -year, -abstract))

setwd("~/git/diversity/data/text_results/h1_results/")
all_cats_counts <- list.files(pattern="h1_diversity_terms_matrix_*") %>% 
  map_df(~read_rds(.)) %>% select(-year, -diversity_cnt, -socdiv_cnt)

combined <- all_diversity_comps %>% 
  left_join(all_cats_counts, by = "fk_pmid") %>% 
  replace(is.na(.), 0) %>% arrange(-socdiv_v7)

setwd("~/git/diversity/data/sensitivity_checks/")
write_csv(combined, "diversity_chk_071121_0813.csv")

counts_by_year <- combined %>% 
  filter(socdiv_v9 == 1) %>% 
  group_by(year) %>% 
  count()

### pluck out random sample on each side of socdiv_v9

socdiv_rand <- combined %>% 
  filter(socdiv_v9 == 1)

socdiv_rand <- sample_n(socdiv_rand, 500)
  
anydiv_rand <- combined %>% 
  filter(socdiv_v9 == 0)

anydiv_rand <- sample_n(anydiv_rand, 500)

div_combined <- bind_rows(socdiv_rand, anydiv_rand)

set.seed(42)
rows <- sample(nrow(div_combined))
div_combined <- div_combined[rows, ]
div_combined_simple <- div_combined %>% select(fk_pmid, abstract)

setwd("~/git/diversity/data/sensitivity_checks/")
write_csv(div_combined, "diversity_tolabel_full_071121.csv")
write_csv(div_combined_simple, "diversity_tolabel_071121.csv")
